# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_scythe504_rbac-finance-processing_internal_database.CategoryTotals:
    properties:
      category:
        type: string
      expense:
        type: string
      income:
        type: string
    type: object
  github_com_scythe504_rbac-finance-processing_internal_database.DashboardSummary:
    properties:
      category_totals:
        items:
          $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_database.CategoryTotals'
        type: array
      grouped_by:
        type: string
      net_balance:
        type: string
      period:
        $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_database.PeriodType'
      recent_activities:
        items:
          $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_database.Record'
        type: array
      total_expense:
        type: string
      total_income:
        type: string
      trends_entry:
        items:
          $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_database.TrendsEntry'
        type: array
    type: object
...
  github_com_scythe504_rbac-finance-processing_internal_database.PeriodType:
    enum:
    - weekly
    - monthly
    - yearly
    - all
    type: string
    x-enum-varnames:
    - PeriodWeekly
    - PeriodMonthly
    - PeriodYearly
    - PeriodAllTime
  github_com_scythe504_rbac-finance-processing_internal_database.Record:
    properties:
      amount:
        type: string
      category:
        type: string
      created_at:
        type: string
      date:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      txn_type:
        $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_database.TxnType'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_scythe504_rbac-finance-processing_internal_database.RecordFilters:
    properties:
      ascending:
        type: boolean
      category:
        type: string
      from:
        type: string
      limit:
        type: integer
      offset:
        type: integer
      show_deleted:
        type: boolean
      to:
        type: string
      txn_type:
        type: string
    type: object
  github_com_scythe504_rbac-finance-processing_internal_database.TrendsEntry:
    properties:
      date:
        type: string
      expense:
        type: string
      income:
        type: string
    type: object
  github_com_scythe504_rbac-finance-processing_internal_database.TxnType:
    enum:
    - income
    - expense
    type: string
    x-enum-varnames:
    - TxnTypeIncome
    - TxnTypeExpense
  github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse:
    properties:
      error:
        example: Internal Server Error
        type: string
    type: object
  github_com_scythe504_rbac-finance-processing_internal_utils.MessageResponse:
    properties:
      message:
        example: success
        type: string
    type: object
  internal_server.AuthResponse:
    properties:
      token:
        example: jwt.token.here
        type: string
      user_id:
        example: "1234"
        type: string
    type: object
  internal_server.CreateRecordRequest:
    properties:
      amount:
        example: "100.50"
        type: string
      category:
        example: Food
        type: string
      date:
        example: "2023-10-27T10:00:00Z"
        type: string
      description:
        example: Lunch at the mall
        type: string
      txn_type:
        enum:
        - income
        - expense
        example: expense
        type: string
    type: object
  internal_server.DashboardResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_database.DashboardSummary'
    type: object
  internal_server.LoginUser:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_server.RecordCreatedResponse:
    properties:
      message:
        example: success
        type: string
      record_id:
        example: 1
        type: integer
    type: object
  internal_server.RecordResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_database.Record'
    type: object
  internal_server.RecordsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_database.Record'
        type: array
      filters:
        $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_database.RecordFilters'
    type: object
  internal_server.RegisterUser:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  internal_server.UpdateRecordRequest:
    properties:
      amount:
        example: "150.00"
        type: string
      category:
        example: Salary
        type: string
      date:
        example: "2023-10-28T09:00:00Z"
        type: string
      description:
        example: Monthly paycheck
        type: string
      txn_type:
        enum:
        - income
        - expense
        example: income
        type: string
    type: object
  internal_server.UserDetailsResponse:
    properties:
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        type: string
      role:
        example: admin
        type: string
      user_id:
        example: "1234"
        type: string
    type: object
info:
  contact: {}
  description: API for managing personal finance with Role-Based Access Control.
  title: RBAC Finance Processing API
  version: 1.0.0
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      operationId: loginUser
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_server.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/internal_server.AuthResponse'
        "400":
          description: Invalid Request Body
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "401":
          description: Invalid Credentials
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      summary: Login a user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new account with a specific role
      operationId: registerUser
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_server.RegisterUser'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/internal_server.AuthResponse'
        "400":
          description: Invalid Request Body / Email Address
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "401":
          description: Self registration is only available for viewer role. Contact
            an admin for desired role.
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "409":
          description: User with email already exists, Please Login
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
  /dashboard:
    get:
      consumes:
      - application/json
      description: Get a summary of income, expenses, trends and recent activities
      operationId: getDashboardSummary
      parameters:
      - description: Summary period (weekly, monthly, yearly, all)
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard summary data
          schema:
            $ref: '#/definitions/internal_server.DashboardResponse'
        "400":
          description: Invalid Query Parameter
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get dashboard summary
      tags:
      - Dashboard
...
  /me:
    get:
      consumes:
      - application/json
      description: Get current logged in user information
      operationId: getUserDetails
      produces:
      - application/json
      responses:
        "200":
          description: User details retrieved
          schema:
            $ref: '#/definitions/internal_server.UserDetailsResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user details
      tags:
      - Users
  /records:
    get:
      consumes:
      - application/json
      description: Get a list of finance records with various filters (Analyst/Admin
        only)
      operationId: getRecords
      parameters:
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by transaction type (income/expense)
        in: query
        name: txn_type
        type: string
      - description: Filter by start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Filter by end date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      - description: Include deleted records
        in: query
        name: show_deleted
        type: boolean
      - description: Sort ascending
        in: query
        name: ascending
        type: boolean
      - description: Limit number of results
        in: query
        name: limit
        type: integer
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of records and filters
          schema:
            $ref: '#/definitions/internal_server.RecordsListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List records
      tags:
      - Records
    post:
      consumes:
      - application/json
      description: Create a new income or expense record (Admin only)
      operationId: createRecord
      parameters:
      - description: Record details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_server.CreateRecordRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Record created
          schema:
            $ref: '#/definitions/internal_server.RecordCreatedResponse'
        "400":
          description: Invalid Request Body
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a record
      tags:
      - Records
  /records/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a record by ID (Admin only)
      operationId: deleteRecord
      parameters:
      - description: Record ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.MessageResponse'
        "400":
          description: Invalid Record ID
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a record
      tags:
      - Records
    get:
      consumes:
      - application/json
      description: Get details of a specific finance record by ID (Analyst/Admin only)
      operationId: getRecord
      parameters:
      - description: Record ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Record details
          schema:
            $ref: '#/definitions/internal_server.RecordResponse'
        "400":
          description: Invalid Record ID
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "404":
          description: Record not found
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a record
      tags:
      - Records
    patch:
      consumes:
      - application/json
      description: Update details of an existing record (Admin only)
      operationId: updateRecord
      parameters:
      - description: Record updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_server.UpdateRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.MessageResponse'
        "400":
          description: Invalid Request Body / ID
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "404":
          description: Record not found
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a record
      tags:
      - Records
  /users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Update the role of a specific user (Admin only)
      operationId: setUserRole
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Target role details
        in: body
        name: body
        required: true
        schema:
          properties:
            target_user_role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.MessageResponse'
        "400":
          description: Invalid Request Body / Role
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set user role
      tags:
      - Users
  /users/{id}/status:
    patch:
      consumes:
      - application/json
      description: Enable or disable a user account (Admin only)
      operationId: toggleUserStatus
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scythe504_rbac-finance-processing_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle user status
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"